[chore] code refactoring
