+++ app.js
@@ -2,7 +2,7 @@
 import express from 'express';
 import logger from 'morgan';
 
-import error from './helpers/errorhandler';
+import error from './helpers/createError';
 import routes from './routes/indexRouter';
 import authenticator from './middleware/authenticator';
 
@@ -15,15 +15,17 @@ app.use(logger('dev'));
 app.use(express.json());
 app.use(express.urlencoded({ extended: false }));
 
-app.use('/api/v1/auth', routes.userRoute);
-app.use('/api/v1/:path', authenticator.verify, routes.generalRoute);
-
-app.use('/api/v1', (req, res) => {
-  res.status(200).json({
-    status: 200,
-    message: 'welcome to questioner, please refer to our API docs for valid requests',
-  });
+app.use('/api/v1', (req, res, next) => {
+  if (req.path === '/') {
+    res.status(200).json({
+      status: 200,
+      message: 'welcome to questioner, please refer to our API docs for valid requests',
+    });
+  } else next();
 });
+
+app.use('/api/v1/auth', routes.userRoute);
+app.use('/api/v1', [authenticator.verify, routes.generalRoute]);
 
 app.use('/:invalid', (req, res) => error(400, res, 'request path invalid, please refer to API documentation'));
